requirements.txt
rdflib >= 4.2.2
PyYAML >= 3.13
